format with single quotes
