+++ src/actions.js
@@ -19,13 +19,3 @@ export const REMOVE_CART_ITEM = 'REMOVE_CART_ITEM';
 export const TOGGLE_CART_ITEM_AMOUNT = 'TOGGLE_CART_ITEM_AMOUNT';
 export const CLEAR_CART = 'CLEAR_CART';
 export const COUNT_CART_TOTALS = 'COUNT_CART_TOTALS';
-export const LOGIN_BEGIN = 'LOGIN_BEGIN';
-export const LOGIN_SUCCESS = 'LOGIN_SUCCESS';
-export const LOGIN_ERROR = 'LOGIN_ERROR';
-export const REGISTER_BEGIN = 'REGISTER_BEGIN';
-export const REGISTER_SUCCESS = 'REGISTER_SUCCESS';
-export const REGISTER_ERROR = 'REGISTER_ERROR';
-export const LOGOUT_BEGIN = 'LOGOUT_BEGIN';
-export const LOGOUT_SUCCESS = 'LOGOUT_SUCCESS';
-export const LOGOUT_ERROR = 'LOGOUT_ERROR';
-export const CURRENT_USER_FETCHED = 'CURRENT_USER_FETCHED';
