+++ src/components/Sidebar.js
@@ -6,7 +6,6 @@ import { FaTimes } from 'react-icons/fa';
 import { links } from '../utils/constants';
 import styled from 'styled-components';
 import CartButtons from './CartButtons';
-import { useUserContext } from '../context/user_context'; //gets the value if user is logged in or not (to display/hide checkout)
 
 const Sidebar = () => {
   const { isSidebarOpen, closeSidebar } = useProductsContext();
