+++ src/components/StripeCheckout.js
@@ -9,7 +9,6 @@ import {
 } from '@stripe/react-stripe-js';
 import axios from 'axios';
 import { useCartContext } from '../context/cart_context';
-import { useUserContext } from '../context/user_context';
 import { formatPrice } from '../utils/helpers';
 import { useHistory } from 'react-router-dom';
