+++ src/index.js
@@ -6,15 +6,12 @@ import App from './App';
 import { ProductsProvider } from './context/products_context';
 import { FilterProvider } from './context/filter_context';
 import { CartProvider } from './context/cart_context';
-import { UserProvider } from './context/user_context';
 import { Auth0Provider } from '@auth0/auth0-react';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
 root.render(
-  <UserProvider>
-    <ProductsProvider>
-      <App />
-    </ProductsProvider>
-  </UserProvider>
+  <ProductsProvider>
+    <App />
+  </ProductsProvider>
 );
