+++ src/utils/constants.js
@@ -40,8 +40,7 @@ export const services = [
 ];
 
 //export const products_url = 'https://course-api.com/react-store-products'
-export const products_url =
-  'https://ecommerce-6kwa.onrender.com/api/v1/products';
+export const products_url = 'https://ecommerce-6kwa.onrender.com/api/v1/products';
 //export const single_product_url = `https://course-api.com/react-store-single-product/api/v1/products?id=`
 export const single_product_url = `https://ecommerce-6kwa.onrender.com/api/v1/products`;
 //point of access to my api
